+++ mathics/builtin/atomic/symbols.py
@@ -8,6 +8,8 @@ or namespace, and can have a variety of type of values and attributes.
 
 import re
 
+from mathics_scanner import is_symbol_name
+
 from mathics.builtin.base import Builtin, PrefixOperator, Test
 from mathics.core.assignment import get_symbol_values
 from mathics.core.atoms import String
@@ -629,7 +631,7 @@ class Names(Builtin):
         "Names[pattern_]"
         headname = pattern.get_head_name()
         if headname == "System`StringExpression":
-            pattern = re.compile(to_regex(pattern, evaluation))
+            pattern = re.compile(to_regex(pattern, show_message=evaluation.messages))
         else:
             pattern = pattern.get_string_value()
 
@@ -714,8 +716,6 @@ class Symbol_(Builtin):
 
     def eval(self, string, evaluation):
         "Symbol[string_String]"
-
-        from mathics.core.parser import is_symbol_name
 
         text = string.value
         if is_symbol_name(text):