+++ CHANGES.rst
@@ -22,6 +22,8 @@ Internals
 ---
 
 * ``eval_abs`` and ``eval_sign`` extracted from ``Abs`` and ``Sign`` and added to ``mathics.eval.arithmetic``.
+* Maximum number of digits allowed in a string set to 7000 and can be adjusted using environment variable
+  ``MATHICS_MAX_STR_DIGITS`` on Python versions that don't adjust automatically (like pyston).
 
 
 Bugs
@@ -102,6 +104,7 @@ New Builtins
 #. ``$PythonImplementation``
 #. ``Accuracy``
 #. ``ClebschGordan``
+#. ``ComplexExpand`` (@yzrun)
 #. ``Curl`` (2-D and 3-D vector forms only)
 #. ``DiscretePlot``
 #. ``Kurtosis``